+++ internal/board/test_cache.go
@@ -1,4 +1,4 @@
-package cache
+package board
 
 import (
 	"sync"
