--- internal/cache/cache.go
@@ -1,7 +1,0 @@
-package cache
-
-type Cache interface {
-	Get(key interface{}) (interface{}, bool)
-	Set(key, value interface{}, cost int64) bool
-	Clear()
-}
