+++ internal/server.go
@@ -9,12 +9,10 @@ import (
 	"os"
 	"time"
 
-	board "github.com/2beens/serjtubincom/internal/board"
-
 	"github.com/2beens/serjtubincom/internal/aerospike"
 	"github.com/2beens/serjtubincom/internal/auth"
 	"github.com/2beens/serjtubincom/internal/blog"
-	"github.com/2beens/serjtubincom/internal/cache"
+	"github.com/2beens/serjtubincom/internal/board"
 	"github.com/2beens/serjtubincom/internal/config"
 	"github.com/2beens/serjtubincom/internal/instrumentation"
 	"github.com/2beens/serjtubincom/internal/middleware"
@@ -74,7 +72,7 @@ func NewServer(
 		return nil, fmt.Errorf("failed to create board aero client: %w", err)
 	}
 
-	boardCache, err := cache.NewBoardCache()
+	boardCache, err := board.NewBoardCache()
 	if err != nil {
 		return nil, fmt.Errorf("failed to create board cache: %w", err)
 	}
